requirements.txt
attrs==22.1.0
awsebcli==3.20.3
bcrypt==3.2.2
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==37.0.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==3.3
importlib-metadata==4.12.0
jmespath==0.10.0
jsonschema==3.2.0
paramiko==2.11.0
pathspec==0.9.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
texttable==1.6.4
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.1.9
websocket-client==0.59.0
zipp==3.8.1
django
djangorestframework-simplejwt
djangorestframework
drf-yasg
pycryptodome